terraform: remove `terraform_modules` generator target type (#15035)
As described in https://github.com/pantsbuild/pants/issues/14787, users have reported inconsistent results and confusion in using the `terraform_modules` target type in the Terraform backend. Moreover, the use of a `terraform_modules` target type is inconsistent with how Pants generates targets for other backends.

This PR makes the Terraform backend's behavior consistent with the other backends by only supporting a `terraform_module` target type. `./pants tailor` will generate it for each Terraform module found.
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
