+++ src/python/pants/backend/experimental/terraform/register.py
@@ -2,19 +2,16 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from pants.backend.python.util_rules.pex import rules as pex_rules
-from pants.backend.terraform import dependency_inference, style, tailor, target_gen, tool
-from pants.backend.terraform.goals import check
+from pants.backend.terraform import dependency_inference, style, tool
+from pants.backend.terraform.goals import check, tailor
 from pants.backend.terraform.lint.tffmt.tffmt import rules as tffmt_rules
-from pants.backend.terraform.target_types import (
-    TerraformModulesGeneratorTarget,
-    TerraformModuleTarget,
-)
+from pants.backend.terraform.target_types import TerraformModuleTarget
 from pants.backend.terraform.target_types import rules as target_types_rules
 from pants.engine.rules import collect_rules
 
 
 def target_types():
-    return [TerraformModuleTarget, TerraformModulesGeneratorTarget]
+    return [TerraformModuleTarget]
 
 
 def rules():
@@ -23,7 +20,6 @@ def rules():
         *check.rules(),
         *dependency_inference.rules(),
         *tailor.rules(),
-        *target_gen.rules(),
         *target_types_rules(),
         *tool.rules(),
         *style.rules(),
